food-ai-agent-api/requirements.txt
# FastAPI & Server
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Settings
pydantic-settings==2.3.4
pydantic[email]==2.7.4

# AI / RAG
anthropic==0.31.2
openai==1.35.13
tiktoken==0.7.0

# Document Processing
PyMuPDF==1.24.7
python-docx==1.1.2

# Utilities
httpx==0.27.0
python-dateutil==2.9.0
apscheduler==3.10.4

# Dev / Test
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
